+++ src/libs/fancy/entities/fancy.entity.ts
@@ -1,8 +1,8 @@
 import { FancyProductStatus } from '@src/api/apps/fancy/enums/fancy-product-status.enum';
 import { ValueOf } from '@src/common/types/common.type';
-import { FancyImageEntity } from '@src/libs/fancy/fancy-images/entity/fancy-image.entity';
-import { FancyOptionEntity } from '@src/libs/fancy/fancy-options/entity/fancy-option.entity';
-import { FancySubOptionEntity } from '@src/libs/fancy/fancy-sub-options/entity/fancy-sub-option.entity';
+import { FancyImageEntity } from '@src/libs/fancy/fancy-images/entities/fancy-image.entity';
+import { FancyOptionEntity } from '@src/libs/fancy/fancy-options/entities/fancy-option.entity';
+import { FancySubOptionEntity } from '@src/libs/fancy/fancy-sub-options/entities/fancy-sub-option.entity';
 import { LookEntity } from '@src/libs/looks/entity/look.entity';
 import { TagEntity } from '@src/libs/tags/entity/tag.entity';
