+++ src/libs/fancy/fancy-images/entities/fancy-image.entity.ts
