+++ src/libs/fancy/fancy-options/entities/fancy-option.entity.ts
